[Chore] Remove build warnings (#1072)
* Remove declarations for methods which no longer exist

* Apply suggestion to silence warning about empty OptionSet

* Mark isSelfConversation as readonly to silence warnign about missing setter

* Change cast to unsigned long so that it matches the string format expression

* Fix incorrect usuage of logging API

+++ Source/Model/ZMManagedObject.m
@@ -741,7 +741,7 @@ static NSString * const KeysForCachedValuesKey = @"ZMKeysForCachedValues";
 
 - (NSString *)nonpersistedObjectIdentifer;
 {
-    return [NSString stringWithFormat:@"Z%tx", (long long) self];
+    return [NSString stringWithFormat:@"Z%tx", (unsigned long) self];
 }
 
 + (instancetype)existingObjectWithNonpersistedObjectIdentifer:(NSString *)identifier inUserSession:(id<ZMManagedObjectContextProvider>)userSession;
